+++ src/Entities/Invoice.php
@@ -180,6 +180,8 @@ final class Invoice extends Entity
     }
 
     /**
+     * @deprecated use {@see getShipping()} instead
+     *
      * @return float
      */
     public function getShippingAmount()
@@ -485,5 +487,33 @@ final class Invoice extends Entity
         return array_map(function ($element) {
             return new InvoiceDiscount($element);
         }, $data);
+    }
+
+    /**
+     * @return Shipping
+     */
+    public function getShipping()
+    {
+        return $this->getAttribute('shipping');
+    }
+
+    /**
+     * @param array|Shipping $value
+     *
+     * @return $this
+     */
+    public function setShipping($value)
+    {
+        return $this->setAttribute('shipping', $value);
+    }
+
+    /**
+     * @param array $data
+     *
+     * @return Shipping
+     */
+    public function createShipping(array $data)
+    {
+        return new Shipping($data);
     }
 }